Quantità decimali Prodotto/Ordine/Spedizioni

+++ Libraries/Nop.Core/Domain/Catalog/Product.cs
@@ -321,7 +321,7 @@ namespace Nop.Core.Domain.Catalog
         /// <summary>
         /// Gets or sets the stock quantity
         /// </summary>
-        public int StockQuantity { get; set; }
+        public decimal StockQuantity { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating whether to display stock availability
@@ -336,7 +336,7 @@ namespace Nop.Core.Domain.Catalog
         /// <summary>
         /// Gets or sets the minimum stock quantity
         /// </summary>
-        public int MinStockQuantity { get; set; }
+        public decimal MinStockQuantity { get; set; }
 
         /// <summary>
         /// Gets or sets the low stock activity identifier
@@ -346,7 +346,7 @@ namespace Nop.Core.Domain.Catalog
         /// <summary>
         /// Gets or sets the quantity when admin should be notified
         /// </summary>
-        public int NotifyAdminForQuantityBelow { get; set; }
+        public decimal NotifyAdminForQuantityBelow { get; set; }
 
         /// <summary>
         /// Gets or sets a value backorder mode identifier
@@ -361,12 +361,12 @@ namespace Nop.Core.Domain.Catalog
         /// <summary>
         /// Gets or sets the order minimum quantity
         /// </summary>
-        public int OrderMinimumQuantity { get; set; }
+        public decimal OrderMinimumQuantity { get; set; }
 
         /// <summary>
         /// Gets or sets the order maximum quantity
         /// </summary>
-        public int OrderMaximumQuantity { get; set; }
+        public decimal OrderMaximumQuantity { get; set; }
 
         /// <summary>
         /// Gets or sets the comma separated list of allowed quantities. null or empty if any quantity is allowed
