+++ Libraries/Nop.Services/Shipping/IShipmentService.cs
@@ -103,7 +103,7 @@ namespace Nop.Services.Shipping
         /// <param name="ignoreShipped">Ignore already shipped shipments</param>
         /// <param name="ignoreDelivered">Ignore already delivered shipments</param>
         /// <returns>Quantity</returns>
-        int GetQuantityInShipments(Product product, int warehouseId,
+        decimal GetQuantityInShipments(Product product, int warehouseId,
             bool ignoreShipped, bool ignoreDelivered);
 
         /// <summary>
