+++ Presentation/Nop.Web/Models/Common/HeaderLinksModel.cs
@@ -8,10 +8,10 @@ namespace Nop.Web.Models.Common
         public string CustomerName { get; set; }
         
         public bool ShoppingCartEnabled { get; set; }
-        public int ShoppingCartItems { get; set; }
+        public decimal ShoppingCartItems { get; set; }
         
         public bool WishlistEnabled { get; set; }
-        public int WishlistItems { get; set; }
+        public decimal WishlistItems { get; set; }
 
         public bool AllowPrivateMessages { get; set; }
         public string UnreadPrivateMessages { get; set; }
