+++ Presentation/Nop.Web/Models/ShoppingCart/ShoppingCartModel.cs
@@ -73,9 +73,9 @@ namespace Nop.Web.Models.ShoppingCart
             public string SubTotal { get; set; }
 
             public string Discount { get; set; }
-            public int? MaximumDiscountedQty { get; set; }
+            public decimal? MaximumDiscountedQty { get; set; }
 
-            public int Quantity { get; set; }
+            public decimal Quantity { get; set; }
             public List<SelectListItem> AllowedQuantities { get; set; }
             
             public string AttributeInfo { get; set; }
